Create panel.js file.
